% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{convert OpenCV to Magick and vice versa}
\usage{
convert(input)
}
\arguments{
\item{input}{either magick-image or opencv-image for conversion}
}
\description{
convert OpenCV to Magick and vice versa
}
\examples{
\dontrun{
 ocvimg <- convert(mag_img)
 magimg <- convert(ocv_img)
}

}
