% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dist_transform}
\alias{dist_transform}
\title{Distance Transform for Binary Files}
\usage{
dist_transform(image, distancetype = 3, masksize = 3, dsttype = 5)
}
\arguments{
\item{image}{An external pointer to an OpenCV binary image}

\item{distancetype}{Type of distance, see cv::DistanceTypes}

\item{masksize}{Size of the distance transform mask, see cv::DistanceTransformMasks.
In case of the DIST_L1 or DIST_C distance type, the
parameter is forced to 3 because a 3×3 mask gives the same result as 5×5 or any
larger aperture.}

\item{dsttype}{Type of output image. It can be CV_8U (0) or CV_32F (5).
Type CV_8U can be used
only for the first variant of the function and distanceType == DIST_L1.}
}
\value{
Output image with calculated distances. It is a 8-bit or 32-bit
floating-point, single-channel image of the same size as src .
}
\description{
Distance Transform for Binary Files
}
