% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{ocv_binarize}
\alias{ocv_binarize}
\title{Binarize grayscale version of image}
\usage{
ocv_binarize(image)
}
\arguments{
\item{image}{A grayscale image}
}
\value{
An external point to a binary image
}
\description{
This is not a robust function, it is only given as an example.
}
