% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/face.R
\name{ocv_face}
\alias{ocv_face}
\title{Face Detection}
\usage{
ocv_face(image)
}
\arguments{
\item{image}{a ocv image object}
}
\description{
Find faces and eyes in a picture.
}
\examples{
\donttest{
unconf <- ocv_read('https://jeroen.github.io/images/unconf18.jpg')
faces <- ocv_face(unconf)
facemask <- ocv_facemask(unconf)
attr(facemask, 'faces')
}
}
