% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{ocv_grayscale}
\alias{ocv_grayscale}
\title{Grayscale version of image}
\usage{
ocv_grayscale(image)
}
\arguments{
\item{image}{A three-channel image}
}
\value{
An external pointer to a grayscale image.
}
\description{
This is not a robust function, it is given as an example.
}
