% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{ocv_text}
\alias{ocv_text}
\title{ocv_text text detection in images using EAST}
\usage{
ocv_text(
  image,
  thrs = 0.5,
  nms = 20,
  width = 320,
  height = 320,
  model,
  draw = FALSE
)
}
\arguments{
\item{image}{Path to input image or video file. Skip this argument to capture
frames from a camera.}

\item{thrs}{Confidence threshold.}

\item{nms}{Non-maximum suppression threshold.}

\item{width}{Preprocess input image by resizing to a specific width. It
should be multiple by 32.}

\item{height}{Preprocess input image by resizing to a specific height. It
should be multiple by 32.}

\item{model}{Path to a binary .pb file contains trained network.}

\item{draw}{Draws visual output to the image.}
}
\description{
OpenCV sample using EAST (Efficient and Accurate Scene Text
Detector) to detect text in images. Requires the EAST pb-model not included
in the package.
}
\examples{
\dontrun{
url <- paste0('https://upload.wikimedia.org/wikipedia/commons/6/6f/',
              'Keep-calm-and-carry-on-scan.jpg')
fl <- ocv_read(url)

ocv_text(image = fl, thrs =  0.7,  nms = 0.3,
         model = "frozen_east_text_detection.pb")
}

}
\references{
https://github.com/opencv/opencv/blob/master/samples/dnn/text_detection.cpp
}
