% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areas.R
\name{opencv-area}
\alias{opencv-area}
\alias{ocv_rectangle}
\title{OpenCV area manipulation}
\usage{
ocv_rectangle(image, x = 0L, y = 0L, width, height)
}
\arguments{
\item{image}{an ocv image object}

\item{x}{horizontal location}

\item{y}{vertical location}

\item{width}{width of the area}

\item{height}{height of the area}
}
\description{
Manipulate image regions
}
\examples{
mona <- ocv_read('https://jeroen.github.io/images/monalisa.jpg')

# Rectangular area
ocv_rectangle(mona, x = 400, y = 300, height = 300, width = 350)
ocv_rectangle(mona, x = 0, y = 100, height = 200)
ocv_rectangle(mona, x = 500, y = 0, width = 75)

ocv_destroy(mona)
}
