% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencv.R
\name{ocv_read}
\alias{ocv_read}
\alias{ocv_write}
\alias{ocv_bitmap}
\alias{ocv_camera}
\alias{print.opencv-image}
\alias{ocv_resize}
\alias{ocv_face}
\alias{ocv_mog2}
\alias{ocv_knn}
\alias{ocv_video}
\title{OpenCV}
\usage{
ocv_read(path)

ocv_write(image, path)

ocv_bitmap(image)

ocv_camera()

\method{print}{opencv-image}(x, ...)

ocv_resize(image, width = 0, height = 0)

ocv_face(image)

ocv_mog2(image)

ocv_knn(image)

ocv_video(filter)
}
\arguments{
\item{x}{passed to [print]}
}
\description{
Utilities to read and write images with OpenCV.s
}
