% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.R
\name{ocv_qr_detect}
\alias{ocv_qr_detect}
\alias{qr_scanner}
\title{Detect QR codes}
\usage{
ocv_qr_detect(image, return_image = FALSE, decoder = c("wechat", "quirc"))

qr_scanner(...)
}
\arguments{
\item{image}{an ocv image object}

\item{decoder}{which decoder implementation to use. The 'wechat' decoder is
the default and generally has better performance and fault-tolerance.}

\item{...}{parameters passed to \link{ocv_qr_detect}}
}
\description{
Detect a QR code from an image. The \code{ocv_qr_detect} function returns the text
content from the QR, and \code{ocv_qr_mask} returns an annotated image with the QR
area highlighted, and text value as an attribute.
}
\examples{
png("test.png")
plot(qrcode::qr_code("This is a test"))
dev.off()
ocv_qrtext(ocv_read('test.png'))
unlink("test.png")
}
