% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencv.R
\name{ocv_read}
\alias{ocv_read}
\alias{ocv_write}
\alias{ocv_bitmap}
\alias{ocv_edges}
\alias{ocv_camera}
\alias{ocv_resize}
\alias{ocv_face}
\alias{ocv_facemask}
\alias{ocv_mog2}
\alias{ocv_knn}
\alias{ocv_hog}
\alias{ocv_blur}
\alias{ocv_sketch}
\alias{ocv_stylize}
\alias{ocv_markers}
\alias{ocv_copyto}
\alias{ocv_display}
\alias{ocv_video}
\title{OpenCV}
\usage{
ocv_read(path)

ocv_write(image, path)

ocv_bitmap(image)

ocv_edges(image)

ocv_camera()

ocv_resize(image, width = 0, height = 0)

ocv_face(image)

ocv_facemask(image)

ocv_mog2(image)

ocv_knn(image)

ocv_hog(image)

ocv_blur(image, ksize = 5)

ocv_sketch(image, color = TRUE)

ocv_stylize(image)

ocv_markers(image)

ocv_copyto(image, target, mask)

ocv_display(image)

ocv_video(filter)
}
\arguments{
\item{path}{image file such as png or jpeg}

\item{image}{a ocv image object}

\item{width}{output width in pixels}

\item{height}{output height in pixels}

\item{ksize}{size of blurring matrix}

\item{color}{true or false}

\item{target}{the output image}

\item{mask}{only copy pixels from the mask}

\item{filter}{an R function that takes and returns an opecv image}

\item{color}{true or false}
}
\description{
Utilities to mess around with OpenCV.
}
\examples{
png('bg.png', width = 1280, height = 720)
par(ask=FALSE)
qplot(speed, dist, data = cars, geom = c("smooth", "point"))
dev.off()
bg <- ocv_read('bg.png')
unlink('pg.png')
ocv_video(function(input){
  mask <- ocv_mog2(input)
  return(ocv_copyto(input, bg, mask))
})

# Overlay face filter
ocv_video(function(input){
  mask <- ocv_facemask(input)
  ocv_copyto(input, bg, mask)
})
}
