% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencv.R
\name{ocv_read}
\alias{ocv_read}
\alias{ocv_write}
\alias{ocv_bitmap}
\alias{ocv_edges}
\alias{ocv_camera}
\alias{ocv_resize}
\alias{ocv_face}
\alias{ocv_mog2}
\alias{ocv_knn}
\alias{ocv_video}
\title{OpenCV}
\usage{
ocv_read(path)

ocv_write(image, path)

ocv_bitmap(image)

ocv_edges(image)

ocv_camera()

ocv_resize(image, width = 0, height = 0)

ocv_face(image)

ocv_mog2(image)

ocv_knn(image)

ocv_video(filter)
}
\arguments{
\item{path}{image file such as png or jpeg}

\item{image}{a ocv image object}

\item{width}{output width in pixels}

\item{height}{output height in pixels}

\item{filter}{an R function that takes and returns an opecv image}
}
\description{
Utilities to read and write images with OpenCV.
}
