% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.R
\name{ocv_qrtext}
\alias{ocv_qrtext}
\alias{ocv_qrmask}
\alias{qr_scanner}
\title{Detect QR codes}
\usage{
ocv_qrtext(image, use_wechat = TRUE)

ocv_qrmask(image, use_wechat = TRUE)

qr_scanner()
}
\arguments{
\item{image}{an ocv image object}

\item{use_wechat}{use the wechat implementation.}
}
\description{
Detect a QR code from an image. The \code{ocv_qrtext} function returns the text
content from the QR, and \code{ocv_qrmask} returns an annotated image with the QR
area highlighted, and the QR text as an attribute.
}
\examples{
png("test.png")
plot(qrcode::qr_code("This is a test"))
dev.off()
ocv_qrtext(ocv_read('test.png'))
unlink("test.png")
}
