% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.R
\name{ocv_qrtext}
\alias{ocv_qrtext}
\alias{ocv_qrmask}
\alias{qr_scanner}
\title{Detect QR codes}
\usage{
ocv_qrtext(image)

ocv_qrmask(image)

qr_scanner()
}
\arguments{
\item{image}{an ocv image object}
}
\description{
Detect a QR code
}
\examples{
png("test.png")
plot(qrcode::qr_code("This is a test"))
dev.off()
ocv_qrtext(ocv_read('test.png'))
unlink("test.png")
\dontrun{
# QR code scanner
ocv_camera(ocv_qrmask)
}
}
