% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.R
\name{ocv_qrtext}
\alias{ocv_qrtext}
\alias{ocv_qrmask}
\title{Detect QR codes}
\usage{
ocv_qrtext(image)

ocv_qrmask(image)
}
\arguments{
\item{image}{an ocv image object}
}
\description{
Detect a QR code
}
\examples{
\dontrun{
# QR code scanner
ocv_camera(ocv_qrmask)
}
}
